Rename Directory + Attempt additional changes
- rename directory (checkstyle fix)
- other refactor tests failed, so this will be the only change
