+++ libtelemetry/src/main/java/com/mapbox/android/telemetry/navigationutilities/AudioTypeChain.java
@@ -1,4 +1,4 @@
-package com.mapbox.android.telemetry.NavigationUtilities;
+package com.mapbox.android.telemetry.navigationutilities;
 
 class AudioTypeChain {
